Consider an extends clause starting with lower case as resolved
It is possible that an extends clause (i.e. what is to the right of the
`extends` keyword) is a fully qualified name and not just a class name.
In this case, we don't want to import anything... but we have no certain
way of telling that. Heuristically, we consider that the clause is fully
qualified if it starts with a lower case letter (as it is the practice
in java to have class names start with an upper case and package name
with a lower case).